fix: Move JSONRPC-specific tests back to AbstractA2AServerTest so that other server implementations that also support JSONRPC can make use of these tests. These tests will be skipped when testing other transports
