+++ reference/jsonrpc/src/test/java/io/a2a/server/apps/quarkus/QuarkusA2AJSONRPCTest.java
@@ -1,47 +1,8 @@
 package io.a2a.server.apps.quarkus;
 
-import static io.restassured.RestAssured.given;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.fail;
-import static org.wildfly.common.Assert.assertNotNull;
-import jakarta.ws.rs.core.MediaType;
-
-import java.net.URI;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.stream.Stream;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-
 import io.a2a.server.apps.common.AbstractA2AServerTest;
-import io.a2a.spec.A2AClientException;
-import io.a2a.spec.InvalidParamsError;
-import io.a2a.spec.InvalidRequestError;
-import io.a2a.spec.JSONParseError;
-import io.a2a.spec.JSONRPCErrorResponse;
-import io.a2a.spec.Message;
-import io.a2a.spec.MessageSendParams;
-import io.a2a.spec.MethodNotFoundError;
-import io.a2a.spec.Part;
-import io.a2a.spec.SendStreamingMessageRequest;
-import io.a2a.spec.SendStreamingMessageResponse;
-import io.a2a.spec.StreamingJSONRPCRequest;
-import io.a2a.spec.Task;
-import io.a2a.spec.TaskQueryParams;
-import io.a2a.spec.TaskState;
-import io.a2a.spec.TextPart;
 import io.a2a.spec.TransportProtocol;
-import io.a2a.util.Utils;
 import io.quarkus.test.junit.QuarkusTest;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
 
 @QuarkusTest
 public class QuarkusA2AJSONRPCTest extends AbstractA2AServerTest {
@@ -58,243 +19,5 @@ public class QuarkusA2AJSONRPCTest extends AbstractA2AServerTest {
     @Override
     protected String getTransportUrl() {
         return "http://localhost:8081";
-    }
-
-    @Test
-    public void testMalformedJSONRPCRequest() {
-        // missing closing bracket
-        String malformedRequest = "{\"jsonrpc\": \"2.0\", \"method\": \"message/send\", \"params\": {\"foo\": \"bar\"}";
-        JSONRPCErrorResponse response = given()
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(malformedRequest)
-                .when()
-                .post("/")
-                .then()
-                .statusCode(200)
-                .extract()
-                .as(JSONRPCErrorResponse.class);
-        assertNotNull(response.getError());
-        assertEquals(new JSONParseError().getCode(), response.getError().getCode());
-    }
-
-    @Test
-    public void testInvalidParamsJSONRPCRequest() {
-        String invalidParamsRequest = """
-            {"jsonrpc": "2.0", "method": "message/send", "params": "not_a_dict", "id": "1"}
-            """;
-        testInvalidParams(invalidParamsRequest);
-
-        invalidParamsRequest = """
-            {"jsonrpc": "2.0", "method": "message/send", "params": {"message": {"parts": "invalid"}}, "id": "1"}
-            """;
-        testInvalidParams(invalidParamsRequest);
-    }
-
-    private void testInvalidParams(String invalidParamsRequest) {
-        JSONRPCErrorResponse response = given()
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(invalidParamsRequest)
-                .when()
-                .post("/")
-                .then()
-                .statusCode(200)
-                .extract()
-                .as(JSONRPCErrorResponse.class);
-        assertNotNull(response.getError());
-        assertEquals(new InvalidParamsError().getCode(), response.getError().getCode());
-        assertEquals("1", response.getId());
-    }
-
-    @Test
-    public void testInvalidJSONRPCRequestMissingJsonrpc() {
-        String invalidRequest = """
-            {
-             "method": "message/send",
-             "params": {}
-            }
-            """;
-        JSONRPCErrorResponse response = given()
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(invalidRequest)
-                .when()
-                .post("/")
-                .then()
-                .statusCode(200)
-                .extract()
-                .as(JSONRPCErrorResponse.class);
-        assertNotNull(response.getError());
-        assertEquals(new InvalidRequestError().getCode(), response.getError().getCode());
-    }
-
-    @Test
-    public void testInvalidJSONRPCRequestMissingMethod() {
-        String invalidRequest = """
-            {"jsonrpc": "2.0", "params": {}}
-            """;
-        JSONRPCErrorResponse response = given()
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(invalidRequest)
-                .when()
-                .post("/")
-                .then()
-                .statusCode(200)
-                .extract()
-                .as(JSONRPCErrorResponse.class);
-        assertNotNull(response.getError());
-        assertEquals(new InvalidRequestError().getCode(), response.getError().getCode());
-    }
-
-    @Test
-    public void testInvalidJSONRPCRequestInvalidId() {
-        String invalidRequest = """
-            {"jsonrpc": "2.0", "method": "message/send", "params": {}, "id": {"bad": "type"}}
-            """;
-        JSONRPCErrorResponse response = given()
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(invalidRequest)
-                .when()
-                .post("/")
-                .then()
-                .statusCode(200)
-                .extract()
-                .as(JSONRPCErrorResponse.class);
-        assertNotNull(response.getError());
-        assertEquals(new InvalidRequestError().getCode(), response.getError().getCode());
-    }
-
-    @Test
-    public void testInvalidJSONRPCRequestNonExistentMethod() {
-        String invalidRequest = """
-            {"jsonrpc": "2.0", "method" : "nonexistent/method", "params": {}}
-            """;
-        JSONRPCErrorResponse response = given()
-                .contentType(MediaType.APPLICATION_JSON)
-                .body(invalidRequest)
-                .when()
-                .post("/")
-                .then()
-                .statusCode(200)
-                .extract()
-                .as(JSONRPCErrorResponse.class);
-        assertNotNull(response.getError());
-        assertEquals(new MethodNotFoundError().getCode(), response.getError().getCode());
-    }
-
-    @Test
-    public void testNonStreamingMethodWithAcceptHeader() throws Exception {
-        testGetTask(MediaType.APPLICATION_JSON);
-    }
-
-    private void testGetTask(String mediaType) throws Exception {
-        saveTaskInTaskStore(MINIMAL_TASK);
-        try {
-            Task response = getClient().getTask(new TaskQueryParams(MINIMAL_TASK.getId()), null);
-            assertEquals("task-123", response.getId());
-            assertEquals("session-xyz", response.getContextId());
-            assertEquals(TaskState.SUBMITTED, response.getStatus().state());
-        } catch (A2AClientException e) {
-            fail("Unexpected exception during getTask: " + e.getMessage(), e);
-        } finally {
-            deleteTaskInTaskStore(MINIMAL_TASK.getId());
-        }
-    }
-
-    @Test
-    public void testStreamingMethodWithAcceptHeader() throws Exception {
-        testSendStreamingMessage(MediaType.SERVER_SENT_EVENTS);
-    }
-
-    @Test
-    public void testSendMessageStreamNewMessageSuccess() throws Exception {
-        testSendStreamingMessage(null);
-    }
-
-    private void testSendStreamingMessage(String mediaType) throws Exception {
-        Message message = new Message.Builder(MESSAGE)
-                .taskId(MINIMAL_TASK.getId())
-                .contextId(MINIMAL_TASK.getContextId())
-                .build();
-        SendStreamingMessageRequest request = new SendStreamingMessageRequest(
-                "1", new MessageSendParams(message, null, null));
-
-        CompletableFuture<HttpResponse<Stream<String>>> responseFuture = initialiseStreamingRequest(request, mediaType);
-
-        CountDownLatch latch = new CountDownLatch(1);
-        AtomicReference<Throwable> errorRef = new AtomicReference<>();
-
-        responseFuture.thenAccept(response -> {
-            if (response.statusCode() != 200) {
-                //errorRef.set(new IllegalStateException("Status code was " + response.statusCode()));
-                throw new IllegalStateException("Status code was " + response.statusCode());
-            }
-            response.body().forEach(line -> {
-                try {
-                    SendStreamingMessageResponse jsonResponse = extractJsonResponseFromSseLine(line);
-                    if (jsonResponse != null) {
-                        assertNull(jsonResponse.getError());
-                        Message messageResponse =  (Message) jsonResponse.getResult();
-                        assertEquals(MESSAGE.getMessageId(), messageResponse.getMessageId());
-                        assertEquals(MESSAGE.getRole(), messageResponse.getRole());
-                        Part<?> part = messageResponse.getParts().get(0);
-                        assertEquals(Part.Kind.TEXT, part.getKind());
-                        assertEquals("test message", ((TextPart) part).getText());
-                        latch.countDown();
-                    }
-                } catch (JsonProcessingException e) {
-                    throw new RuntimeException(e);
-                }
-            });
-        }).exceptionally(t -> {
-            if (!isStreamClosedError(t)) {
-                errorRef.set(t);
-            }
-            latch.countDown();
-            return null;
-        });
-
-
-        boolean dataRead = latch.await(20, TimeUnit.SECONDS);
-        Assertions.assertTrue(dataRead);
-        Assertions.assertNull(errorRef.get());
-
-    }
-
-    private CompletableFuture<HttpResponse<Stream<String>>> initialiseStreamingRequest(
-            StreamingJSONRPCRequest<?> request, String mediaType) throws Exception {
-
-        // Create the client
-        HttpClient client = HttpClient.newBuilder()
-                .version(HttpClient.Version.HTTP_2)
-                .build();
-
-        // Create the request
-        HttpRequest.Builder builder = HttpRequest.newBuilder()
-                .uri(URI.create("http://localhost:" + serverPort + "/"))
-                .POST(HttpRequest.BodyPublishers.ofString(Utils.OBJECT_MAPPER.writeValueAsString(request)))
-                .header("Content-Type", APPLICATION_JSON);
-        if (mediaType != null) {
-            builder.header("Accept", mediaType);
-        }
-        HttpRequest httpRequest = builder.build();
-
-
-        // Send request async and return the CompletableFuture
-        return client.sendAsync(httpRequest, HttpResponse.BodyHandlers.ofLines());
-    }
-
-    private SendStreamingMessageResponse extractJsonResponseFromSseLine(String line) throws JsonProcessingException {
-        line = extractSseData(line);
-        if (line != null) {
-            return Utils.OBJECT_MAPPER.readValue(line, SendStreamingMessageResponse.class);
-        }
-        return null;
-    }
-
-    private static String extractSseData(String line) {
-        if (line.startsWith("data:")) {
-            line =  line.substring(5).trim();
-            return line;
-        }
-        return null;
     }
 }
