+++ tests/server-common/src/test/java/io/a2a/server/apps/common/AbstractA2AServerTest.java
@@ -1,12 +1,16 @@
 package io.a2a.server.apps.common;
 
+import static io.restassured.RestAssured.given;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.fail;
+import static org.junit.jupiter.api.Assumptions.assumeTrue;
 import static org.wildfly.common.Assert.assertNotNull;
 import static org.wildfly.common.Assert.assertTrue;
+
+import jakarta.ws.rs.core.MediaType;
 
 import java.io.EOFException;
 import java.io.IOException;
@@ -25,6 +29,9 @@ import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
+import java.util.stream.Stream;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
 
 import io.a2a.client.Client;
 import io.a2a.client.ClientEvent;
@@ -41,10 +48,19 @@ import io.a2a.spec.Artifact;
 import io.a2a.spec.DeleteTaskPushNotificationConfigParams;
 import io.a2a.spec.Event;
 import io.a2a.spec.GetTaskPushNotificationConfigParams;
+import io.a2a.spec.InvalidParamsError;
+import io.a2a.spec.InvalidRequestError;
+import io.a2a.spec.JSONParseError;
+import io.a2a.spec.JSONRPCErrorResponse;
 import io.a2a.spec.ListTaskPushNotificationConfigParams;
 import io.a2a.spec.Message;
+import io.a2a.spec.MessageSendParams;
+import io.a2a.spec.MethodNotFoundError;
 import io.a2a.spec.Part;
 import io.a2a.spec.PushNotificationConfig;
+import io.a2a.spec.SendStreamingMessageRequest;
+import io.a2a.spec.SendStreamingMessageResponse;
+import io.a2a.spec.StreamingJSONRPCRequest;
 import io.a2a.spec.Task;
 import io.a2a.spec.TaskArtifactUpdateEvent;
 import io.a2a.spec.TaskIdParams;
@@ -55,9 +71,11 @@ import io.a2a.spec.TaskState;
 import io.a2a.spec.TaskStatus;
 import io.a2a.spec.TaskStatusUpdateEvent;
 import io.a2a.spec.TextPart;
+import io.a2a.spec.TransportProtocol;
 import io.a2a.spec.UnsupportedOperationError;
 import io.a2a.util.Utils;
 
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
 
@@ -781,6 +799,265 @@ public abstract class AbstractA2AServerTest {
             deletePushNotificationConfigInStore(MINIMAL_TASK.getId(), MINIMAL_TASK.getId());
             deleteTaskInTaskStore(MINIMAL_TASK.getId());
         }
+    }
+
+    @Test
+    public void testMalformedJSONRPCRequest() {
+        // skip this test for non-JSONRPC transports
+        assumeTrue(TransportProtocol.JSONRPC.asString().equals(getTransportProtocol()),
+                "JSONRPC-specific test");
+        
+        // missing closing bracket
+        String malformedRequest = "{\"jsonrpc\": \"2.0\", \"method\": \"message/send\", \"params\": {\"foo\": \"bar\"}";
+        JSONRPCErrorResponse response = given()
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(malformedRequest)
+                .when()
+                .post("/")
+                .then()
+                .statusCode(200)
+                .extract()
+                .as(JSONRPCErrorResponse.class);
+        assertNotNull(response.getError());
+        assertEquals(new JSONParseError().getCode(), response.getError().getCode());
+    }
+
+    @Test
+    public void testInvalidParamsJSONRPCRequest() {
+        // skip this test for non-JSONRPC transports
+        assumeTrue(TransportProtocol.JSONRPC.asString().equals(getTransportProtocol()),
+                "JSONRPC-specific test");
+        
+        String invalidParamsRequest = """
+            {"jsonrpc": "2.0", "method": "message/send", "params": "not_a_dict", "id": "1"}
+            """;
+        testInvalidParams(invalidParamsRequest);
+
+        invalidParamsRequest = """
+            {"jsonrpc": "2.0", "method": "message/send", "params": {"message": {"parts": "invalid"}}, "id": "1"}
+            """;
+        testInvalidParams(invalidParamsRequest);
+    }
+
+    private void testInvalidParams(String invalidParamsRequest) {
+        JSONRPCErrorResponse response = given()
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(invalidParamsRequest)
+                .when()
+                .post("/")
+                .then()
+                .statusCode(200)
+                .extract()
+                .as(JSONRPCErrorResponse.class);
+        assertNotNull(response.getError());
+        assertEquals(new InvalidParamsError().getCode(), response.getError().getCode());
+        assertEquals("1", response.getId());
+    }
+
+    @Test
+    public void testInvalidJSONRPCRequestMissingJsonrpc() {
+        // skip this test for non-JSONRPC transports
+        assumeTrue(TransportProtocol.JSONRPC.asString().equals(getTransportProtocol()),
+                "JSONRPC-specific test");
+        
+        String invalidRequest = """
+            {
+             "method": "message/send",
+             "params": {}
+            }
+            """;
+        JSONRPCErrorResponse response = given()
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(invalidRequest)
+                .when()
+                .post("/")
+                .then()
+                .statusCode(200)
+                .extract()
+                .as(JSONRPCErrorResponse.class);
+        assertNotNull(response.getError());
+        assertEquals(new InvalidRequestError().getCode(), response.getError().getCode());
+    }
+
+    @Test
+    public void testInvalidJSONRPCRequestMissingMethod() {
+        // skip this test for non-JSONRPC transports
+        assumeTrue(TransportProtocol.JSONRPC.asString().equals(getTransportProtocol()),
+                "JSONRPC-specific test");
+        
+        String invalidRequest = """
+            {"jsonrpc": "2.0", "params": {}}
+            """;
+        JSONRPCErrorResponse response = given()
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(invalidRequest)
+                .when()
+                .post("/")
+                .then()
+                .statusCode(200)
+                .extract()
+                .as(JSONRPCErrorResponse.class);
+        assertNotNull(response.getError());
+        assertEquals(new InvalidRequestError().getCode(), response.getError().getCode());
+    }
+
+    @Test
+    public void testInvalidJSONRPCRequestInvalidId() {
+        // skip this test for non-JSONRPC transports
+        assumeTrue(TransportProtocol.JSONRPC.asString().equals(getTransportProtocol()),
+                "JSONRPC-specific test");
+        
+        String invalidRequest = """
+            {"jsonrpc": "2.0", "method": "message/send", "params": {}, "id": {"bad": "type"}}
+            """;
+        JSONRPCErrorResponse response = given()
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(invalidRequest)
+                .when()
+                .post("/")
+                .then()
+                .statusCode(200)
+                .extract()
+                .as(JSONRPCErrorResponse.class);
+        assertNotNull(response.getError());
+        assertEquals(new InvalidRequestError().getCode(), response.getError().getCode());
+    }
+
+    @Test
+    public void testInvalidJSONRPCRequestNonExistentMethod() {
+        // skip this test for non-JSONRPC transports
+        assumeTrue(TransportProtocol.JSONRPC.asString().equals(getTransportProtocol()),
+                "JSONRPC-specific test");
+        
+        String invalidRequest = """
+            {"jsonrpc": "2.0", "method" : "nonexistent/method", "params": {}}
+            """;
+        JSONRPCErrorResponse response = given()
+                .contentType(MediaType.APPLICATION_JSON)
+                .body(invalidRequest)
+                .when()
+                .post("/")
+                .then()
+                .statusCode(200)
+                .extract()
+                .as(JSONRPCErrorResponse.class);
+        assertNotNull(response.getError());
+        assertEquals(new MethodNotFoundError().getCode(), response.getError().getCode());
+    }
+
+    @Test
+    public void testNonStreamingMethodWithAcceptHeader() throws Exception {
+        // skip this test for non-JSONRPC transports
+        assumeTrue(TransportProtocol.JSONRPC.asString().equals(getTransportProtocol()),
+                "JSONRPC-specific test");
+        testGetTask(MediaType.APPLICATION_JSON);
+    }
+
+    @Test
+    public void testStreamingMethodWithAcceptHeader() throws Exception {
+        // skip this test for non-JSONRPC transports
+        assumeTrue(TransportProtocol.JSONRPC.asString().equals(getTransportProtocol()),
+                "JSONRPC-specific test");
+        
+        testSendStreamingMessage(MediaType.SERVER_SENT_EVENTS);
+    }
+
+    @Test
+    public void testSendMessageStreamNewMessageSuccess() throws Exception {
+        // skip this test for non-JSONRPC transports
+        assumeTrue(TransportProtocol.JSONRPC.asString().equals(getTransportProtocol()),
+                "JSONRPC-specific test");
+        
+        testSendStreamingMessage(null);
+    }
+
+    private void testSendStreamingMessage(String mediaType) throws Exception {
+        Message message = new Message.Builder(MESSAGE)
+                .taskId(MINIMAL_TASK.getId())
+                .contextId(MINIMAL_TASK.getContextId())
+                .build();
+        SendStreamingMessageRequest request = new SendStreamingMessageRequest(
+                "1", new MessageSendParams(message, null, null));
+
+        CompletableFuture<HttpResponse<Stream<String>>> responseFuture = initialiseStreamingRequest(request, mediaType);
+
+        CountDownLatch latch = new CountDownLatch(1);
+        AtomicReference<Throwable> errorRef = new AtomicReference<>();
+
+        responseFuture.thenAccept(response -> {
+            if (response.statusCode() != 200) {
+                //errorRef.set(new IllegalStateException("Status code was " + response.statusCode()));
+                throw new IllegalStateException("Status code was " + response.statusCode());
+            }
+            response.body().forEach(line -> {
+                try {
+                    SendStreamingMessageResponse jsonResponse = extractJsonResponseFromSseLine(line);
+                    if (jsonResponse != null) {
+                        assertNull(jsonResponse.getError());
+                        Message messageResponse =  (Message) jsonResponse.getResult();
+                        assertEquals(MESSAGE.getMessageId(), messageResponse.getMessageId());
+                        assertEquals(MESSAGE.getRole(), messageResponse.getRole());
+                        Part<?> part = messageResponse.getParts().get(0);
+                        assertEquals(Part.Kind.TEXT, part.getKind());
+                        assertEquals("test message", ((TextPart) part).getText());
+                        latch.countDown();
+                    }
+                } catch (JsonProcessingException e) {
+                    throw new RuntimeException(e);
+                }
+            });
+        }).exceptionally(t -> {
+            if (!isStreamClosedError(t)) {
+                errorRef.set(t);
+            }
+            latch.countDown();
+            return null;
+        });
+
+
+        boolean dataRead = latch.await(20, TimeUnit.SECONDS);
+        Assertions.assertTrue(dataRead);
+        Assertions.assertNull(errorRef.get());
+
+    }
+
+    private CompletableFuture<HttpResponse<Stream<String>>> initialiseStreamingRequest(
+            StreamingJSONRPCRequest<?> request, String mediaType) throws Exception {
+
+        // Create the client
+        HttpClient client = HttpClient.newBuilder()
+                .version(HttpClient.Version.HTTP_2)
+                .build();
+
+        // Create the request
+        HttpRequest.Builder builder = HttpRequest.newBuilder()
+                .uri(URI.create("http://localhost:" + serverPort + "/"))
+                .POST(HttpRequest.BodyPublishers.ofString(Utils.OBJECT_MAPPER.writeValueAsString(request)))
+                .header("Content-Type", APPLICATION_JSON);
+        if (mediaType != null) {
+            builder.header("Accept", mediaType);
+        }
+        HttpRequest httpRequest = builder.build();
+
+
+        // Send request async and return the CompletableFuture
+        return client.sendAsync(httpRequest, HttpResponse.BodyHandlers.ofLines());
+    }
+
+    private SendStreamingMessageResponse extractJsonResponseFromSseLine(String line) throws JsonProcessingException {
+        line = extractSseData(line);
+        if (line != null) {
+            return Utils.OBJECT_MAPPER.readValue(line, SendStreamingMessageResponse.class);
+        }
+        return null;
+    }
+
+    private static String extractSseData(String line) {
+        if (line.startsWith("data:")) {
+            line =  line.substring(5).trim();
+            return line;
+        }
+        return null;
     }
 
     protected boolean isStreamClosedError(Throwable throwable) {
